[Update] Restructured test files.
