+++ src/components/navigation/index.js
@@ -1,3 +1,7 @@
 import _Navigation from './navigation';
+import _style from './style';
 
-export const Navigation = _Navigation;
+export default _Navigation;
+
+export const Navigation = _Navigation,
+  style = _style;
