+++ src/components/navigation/navigation.js
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react';
 import { Link } from 'react-router';
 import { connect } from 'react-redux';
-import style from './style/navigation';
+import style from './style';
 
 class Navigation extends Component {
   renderSignInOrOut() {
