+++ src/components/title/index.js
@@ -1,5 +1,5 @@
 import title from './title';
-import _style from './style/title';
+import _style from './style';
 
 export default title;
 export const style = _style;
