+++ src/components/title/test/mount-spec.js
@@ -1,18 +1,9 @@
 import { expect } from 'helpers/test-helper';
 
-import App, { styles } from 'components/app';
 import Title from 'components/title';
 
-describe('#Initial tests setup', function() {
-  it('-Render App component', function() {
-    const container = document.getElementById('app'),
-      _node = renderComponent(App, container, {}),
-      className = styles.app;
-
-    expect(_node.className).to.equal(className);
-  });
-
-  it('-Render Title', function() {
+describe('#Title component', function() {
+  it('-Classname is mapped', function() {
     const container = document.getElementById('app'),
       _node = renderComponent(Title, container, {
         title: 'App Title'
